horseshitbot_robot_web_clean/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
pymodbus==3.7.4
python-dotenv==1.0.1
pyserial==3.5
